TS header seach function now gives feedback if the timestamp searched lies outside the range of the file

+++ pi_Cpp/dataStorage/MainData.h
@@ -43,7 +43,7 @@ public:
    * goes through the data inbetween start and stop, extracts the time from the data and puts
    * in the time array executes a given function on the data for every datapoint and puts that
    * in the float array. Returns the length of the float array */
-  uint16_t fetchData(uint32_t startT, uint32_t stopT, uint32_t x[], float y[],
+  int fetchData(uint32_t startT, uint32_t stopT, uint32_t x[], float y[],
                      float (*func)(int orgIdx_B, int blockIdx_B, uint8_t[MAXBLOCKSIZE],
                      int extraParams[4]), int extraParams[4]);
